+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased][unreleased]
+### Added
+- `Knock.token_lifetime` configuration variable
+- `Knock.token_secret_signature_key` configuration variable
+
 ## [1.0.0] - 2015-14-07
 ### Fixed
 - Replaced all remaining references to deprecated gem name (Simsim)